+++ eslint.config.mjs
@@ -29,6 +29,7 @@ export default defineConfig(
 			"@typescript-eslint/consistent-type-imports": "error",
 			"@typescript-eslint/no-floating-promises": "error",
 			"@typescript-eslint/no-misused-promises": "error",
+			"obsidianmd/commands/no-default-hotkeys": "off",
 		},
 	},
 	{